major: any plugins removed
See it's been a long time I have been using this plugins, But now it seems like some plugins are causing performance issues in neovim. So this commit changes that. Not only performance issues but many plugins that I don't use any more are also being removed.

The list of those plugins are as follows:
1. nvim-telescope/telescope-fzy-native.nvim
2. Xuyuanp/nerdtree-git-plugin
3. folke/trouble.nvim
4. folke/todo-comments.nvim
5. itchyny/vim-cursorword
6. kana/vim-repeat
7. and more

also there config files are removed.